Test Cases - commit after running tests, run tests --without_serve

+++ frappe/cli.py
@@ -155,6 +155,8 @@ def setup_test(parser):
 		help="Run command for specified module")
 	parser.add_argument("--tests", metavar="TEST FUNCTION", nargs="*",
 		help="Run one or more specific test functions")
+	parser.add_argument("--without_serve", default=False, action="store_true",
+		help="Run tests assuming serve is executed manually on port 8888")
 
 def setup_utilities(parser):
 	# serving
@@ -701,16 +703,18 @@ def smtp_debug_server():
 	os.execv(python, [python, '-m', "smtpd", "-n", "-c", "DebuggingServer", "localhost:25"])
 
 @cmd
-def run_tests(app=None, module=None, doctype=None, verbose=False, tests=()):
+def run_tests(app=None, module=None, doctype=None, verbose=False, tests=(), without_serve=False):
 	import frappe.test_runner
 	from frappe.utils import sel
 
 	frappe.local.localhost = "http://localhost:8888"
-	pipe = subprocess.Popen(["frappe", "--serve", "--port", "8888"], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-	while not pipe.stderr.readline():
-		time.sleep(0.5)
-	if verbose:
-		print "Test server started"
+	if not without_serve:
+		pipe = subprocess.Popen(["frappe", frappe.local.site, "--serve", "--port", "8888"],
+			stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+		while not pipe.stderr.readline():
+			time.sleep(0.5)
+		if verbose:
+			print "Test server started"
 
 	sel.start(verbose)
 
@@ -721,7 +725,9 @@ def run_tests(app=None, module=None, doctype=None, verbose=False, tests=()):
 		if len(ret.failures) == 0 and len(ret.errors) == 0:
 			ret = 0
 	finally:
-		pipe.terminate()
+		if not without_serve:
+			pipe.terminate()
+
 		sel.close()
 
 	return ret
